% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{minionverse_deps}
\alias{minionverse_deps}
\title{List all minionverse dependencies}
\usage{
minionverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
minionverse packages.}
}
\description{
List all minionverse dependencies
}
