% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{minionverse_packages}
\alias{minionverse_packages}
\title{List all packages in the minionverse}
\usage{
minionverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include minionverse in the list?}
}
\description{
List all packages in the minionverse
}
\examples{
minionverse_packages()
}
