% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{minionverse_update}
\alias{minionverse_update}
\title{Update minionverse packages}
\usage{
minionverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
minionverse packages.}
}
\description{
This will check to see if all minionverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
minionverse_update()
}
}
